Try to programmatically link A-P entries to AniDB.
Currently I'm being lazy, and just using some known information (parsing out
data on an anime is the point of this project, after all) to give A-P's
advanced search a reasonable chance of finding only one result.  In the future,
we should perhaps try and figure out the proper result when we have multiple
returned, but that's Too Much Work until I see what kind of success this simple
method has.