Improve regex match for pickier databases.
For reasons I'm not sure of, I decided earlier it would be a good idea to
include a trailing plus sign in the episode count, even though the episode
count is supposed to be an integer.  SQLite was apparently ok with this,
trimming it off, but MySQL was rather unhappy.